fix: moved localStorageHelper.ts to utils folder
